Move to the standard package used by cobra-cli
