+++ src/go/cmd/doc.go
@@ -1,2 +1,2 @@
 // The Command Line Interface (CLI) for Marmot, broken up into packages for each top-level command.
-package cli
+package cmd
